Implemented stepper driver and hardware reset
Implemented a basic pulsed stepper driver on the arduino uno to more
accurately replicated the use of a brushless DC motor in the final
project. This uses timer generated interrupts and direct port access as
opposed to the Arduino delay and write functions used by other
libraries.
Also implemented a hardware reset command that uses wires attached to
the reset pin on Arduino boards. The Mega first resets the two Unos,
then itself.