Rewrote command system, servo on mega
In order to solve issues with bi-directional communication using an Xbee
I moved to a system whereby a command from the PC sets a given amount of
data points to return. This requires very precise motor control and
therefore a separate system on I2C produces too much lag. I therefore
moved the servo driver functions to the mega sketch and modified the
python script to send commands on receiving an acknowledgment from the
arduino.