Added micro-stepping, background ADC, and GUI
Implemented microstepping on the mega using three pins. Takes a scale
(1,2,5,8,16,32) and converts it to a value from 0 to 5 which is then
used to set the appropriate pin combinations.
Also switched from using Analog Read to setting up the ADC to run in the
background and fire an interrupt when done. The ADMUX register is used
to iterate through three input pins (A5-A7, I think I may have fried
some of the others) before starting the next conversion. This means that
the sketch never stops and waits for a conversion to finish, the read
values are simply updated each interrupt.
Finally also implemented a simple control GUI using PyQT4. This can send
commands and receive data and also makes use of the threading library to
prevent the read functions blocking the rest of the program.